Extract DAL interface for report model into a separate report_dal.py file
